Remove backlog item reference from monitoring page
As this turned out to be a feature of not much value add (BDBA + GitHub
reference is also shown on compliance cells) + is rather specific for
these extensions (i.e. there will be no reference for ClamAV or SAST
or ...) + required some special handling (-> merging of BDBA base url in
scan config on delivery-service side), drop this code path.