Don't display snackbar for failed service extension fetch
This is especially relevant in case an unauthenticated user tries to
access the Delivery-Dashboard, as the service extension fetch will be
done anyways and results in a 401 error, ultimately causing the user to
be redirected to the login page. However, the user should not be
confused by an unnecessary snackbar stating failure of service extension
fetch.