feat(bom): Deduplicate component-versions
An OCM tree can contain multiple references of the very same OCM
component-version. Avoid noise and confusion by deduplicating.
Extend "Referenced By" section to display every single reference.

Note:
As an individual entry in BoM view may not be unique anymore
(component-version is referenced multiple items), it is not obvious with
reference to use for updating the "OCM path" in navigation bar.

Signed-off-by: Philipp Heil (zkdev) <philipp.heil@sap.com>